requirements.txt
flask
langchain==0.2.17
langchain-openai==0.1.25
networkx
matplotlib
python-dotenv
